+++ common/src/main/kotlin/com/cobblemon/mod/common/pokemon/feature/SeasonFeature.kt
@@ -38,9 +38,13 @@ const val SEASON = "season"
 
 object SeasonFeatureHandler {
     fun updateSeason(pokemon: Pokemon, world: WorldAccess, pos: BlockPos) {
+        updateSeason(pokemon, Cobblemon.seasonResolver(world, pos))
+    }
+
+    fun updateSeason(pokemon: Pokemon, season: CobblemonSeason?) {
         val feature = pokemon.getFeature<StringSpeciesFeature>(SEASON) ?: return
         val currentSeason = feature.value
-        val newSeason = Cobblemon.seasonResolver(world, pos)?.name?.lowercase()
+        val newSeason = season?.name?.lowercase()
         if (currentSeason != newSeason && newSeason != null) {
             feature.value = newSeason
             pokemon.updateAspects()