+++ CHANGELOG.md
@@ -116,6 +116,7 @@
 - Fixed Pokémon layers not rendering when a Pokémon is on your shoulder.
 - Fixed Caterpie and Weedle not moving or looking at players. That was meant to be Metapod and Kakuna, how embarrassing.
 - Fixed Pokémon not carrying over the correct equivalent original ability when evolving from stages that only had one ability.
+- Fixed Deerling and Sawsbuck not spawning with the correct season.
 
 ### Developer
 - Reworked CatchRateModifier, as such, existing implementations need to be updated.